requirements.txt
numpy>=1.21.0
pytest>=6.2.5
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
torch>=1.9.0  # PyTorch for tensor operations and autograd
torchvision>=0.10.0  # For additional PyTorch utilities
torchaudio>=0.9.0  # For additional PyTorch utilities
deuces>=0.1  # For fast hand evaluation
python-dotenv>=0.19.0
pytest-asyncio>=0.15.1
